pyproject.toml
[project]
name = "breathe-flow-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage>=3.10.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
